Minor fix to format and handling

+++ Windows/ProactiveRemediations/Remediation-MitigateCVE202230190.ps1
@@ -9,55 +9,59 @@ $RegistryRoot = $null
 
 # Figure out if we have an existing PowerShell Registry Provider mapping
 $paramGetPSDrive = @{
-    ErrorAction   = $SCT
-    WarningAction = $SCT
+   ErrorAction   = $SCT
+   WarningAction = $SCT
 }
 $RegistryRoot = ((Get-PSDrive @paramGetPSDrive | Where-Object {
-        $PSItem.Root -eq 'HKEY_CLASSES_ROOT'
-        }).Name)
+         $PSItem.Root -eq 'HKEY_CLASSES_ROOT'
+      }).Name)
 $paramGetPSDrive = $null
 
 if (-not ($RegistryRoot))
 {
-    # PowerShell Registry Provider
-    $paramNewPSDrive = @{
-        PSProvider  = 'registry'
-        Root        = 'HKEY_CLASSES_ROOT'
-        Name        = 'HKCR'
-        ErrorAction = $STP
-    }
-    $RegistryRoot = ((New-PSDrive @paramNewPSDrive).Name)
-    $paramNewPSDrive = $null
+   # PowerShell Registry Provider
+   $paramNewPSDrive = @{
+      PSProvider  = 'registry'
+      Root        = 'HKEY_CLASSES_ROOT'
+      Name        = 'HKCR'
+      ErrorAction = $STP
+   }
+   $RegistryRoot = ((New-PSDrive @paramNewPSDrive).Name)
+   $paramNewPSDrive = $null
 }
 
 try
 {
-    # Guidance for CVE-2022-30190 Microsoft Support Diagnostic Tool Vulnerability
-    $paramRemoveItem = @{
-        Path        = ('{0}:\ms-msdt' -f $RegistryRoot)
-        Force       = $true
-        Recurse     = $true
-        ErrorAction = $STP
-    }
-    $null = (Remove-Item @paramRemoveItem)
-    $paramRemoveItem = $null
+   If (Get-Item -Path ('{0}:\ms-msdt' -f $RegistryRoot) -ErrorAction $SCT)
+   {
+      # Guidance for CVE-2022-30190 Microsoft Support Diagnostic Tool Vulnerability
+      $paramRemoveItem = @{
+         Path          = ('{0}:\ms-msdt' -f $RegistryRoot)
+         Force         = $true
+         Recurse       = $true
+         ErrorAction   = $STP
+         WarningAction = $SCT
+      }
+      $null = (Remove-Item @paramRemoveItem)
+      $paramRemoveItem = $null
+   }
 }
 catch
 {
-    [Management.Automation.ErrorRecord]$e = $_
+   [Management.Automation.ErrorRecord]$e = $_
 
-    $info = [PSCustomObject]@{
-        Exception = $e.Exception.Message
-        Reason    = $e.CategoryInfo.Reason
-        Target    = $e.CategoryInfo.TargetName
-        Script    = $e.InvocationInfo.ScriptName
-        Line      = $e.InvocationInfo.ScriptLineNumber
-        Column    = $e.InvocationInfo.OffsetInLine
-    }
+   $info = [PSCustomObject]@{
+      Exception = $e.Exception.Message
+      Reason    = $e.CategoryInfo.Reason
+      Target    = $e.CategoryInfo.TargetName
+      Script    = $e.InvocationInfo.ScriptName
+      Line      = $e.InvocationInfo.ScriptLineNumber
+      Column    = $e.InvocationInfo.OffsetInLine
+   }
 
-    Write-Warning -Message $info -WarningAction $STP
+   Write-Warning -Message $info -WarningAction $STP
 
-    exit 1
+   exit 1
 }
 
 return $true
